% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alepr.R
\name{alepr_projeto.data.frame}
\alias{alepr_projeto.data.frame}
\title{Método para extrair dados de um data frame}
\usage{
\method{alepr_projeto}{data.frame}(
  obj,
  numero = NULL,
  ano = NULL,
  codigoTipoNormaLegal = NULL,
  .report = TRUE
)
}
\arguments{
\item{obj}{Um data frame contendo os dados de entrada.}

\item{numero}{Coluna do data frame contendo números de norma legal.}

\item{ano}{Coluna do data frame contendo anos de norma legal.}

\item{codigoTipoNormaLegal}{Coluna do data frame contendo códigos de tipo de norma legal.}

\item{.report}{Se TRUE, mostra informações sobre a extração de dados.}
}
\value{
Um data frame contendo os dados extraídos da API.
}
\description{
Este método extrai dados da API da Assembleia Legislativa do Estado do Paraná com base em um data frame.
}
\examples{
dados <- data.frame(numero = c(1, 2, 3), ano = c(2020, 2021, 2022), codigoTipoNormaLegal = c("A", "B", "C"))
alepr_projeto.data.frame(dados)
}
