% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alepr.R
\name{alepr_projeto}
\alias{alepr_projeto}
\title{Função principal alepr_projeto}
\usage{
alepr_projeto(...)

## s3 method for class 'data.frame'
alepr_projeto(
 obj,
 numero = NULL,
 ano = NULL,
 codigoTipoNormaLegal = NULL,
 .report = TRUE)
}
\arguments{
\item{...}{Argumentos adicionais que serão passados para métodos específicos.}

\item{.report}{Se TRUE, mostra informações sobre a extração de dados.}
}
\description{
Esta função extrai dados da API da Assembleia Legislativa do Estado do Paraná.
}
\details{
\subsection{Método data.frame}{

Argumento obrigatório \code{obj} que deve ser um data.frame. E os argumentos opcionais: \code{numero}, \code{ano} e \code{codigoTipoNormaLegal}, que devem conter as colunas do data frame com os respectivos valores. Se você não quiser um resumo ao final, use \code{.report = FALSE}.
}

\subsection{Método numeric}{
}
}
\examples{
\dontrun{
alepr_projeto()
}
}
\seealso{
Other legislativo: 
\code{\link{alers_projeto}()}
}
\concept{legislativo}
