% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alers.R
\name{alers_projeto}
\alias{alers_projeto}
\title{Buscar informações sobre proposição legislativa na Assembleia Legislativa do RS}
\usage{
alers_projeto(norma, ano, .reportar = TRUE)
}
\arguments{
\item{norma}{Número da norma.}

\item{ano}{Ano da proposição.}

\item{.reportar}{Se TRUE, exibe uma mensagem informando sobre a proposição buscada. O padrão é TRUE.}
}
\value{
Um data frame contendo informações sobre a proposição legislativa, incluindo o número da norma, a proposição,
o proponente, o assunto, a ementa e a URL da proposição.
}
\description{
Esta função busca informações sobre uma proposição legislativa na Assembleia Legislativa do Rio Grande do Sul (ALERS),
dado o número da norma e o ano de sua proposição.
}
\examples{
\dontrun{
# Buscar informações sobre uma proposição legislativa
alers_projeto(norma = 10098, ano = 1994)
}
}
\seealso{
Other legislativo: 
\code{\link{alepr_projeto}()}
}
\concept{legislativo}
