% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_4grams}
\alias{dic_4grams}
\title{Dicionários de expressões com 4 palavras}
\format{
\subsection{\code{dic_4grams}}{

Um data frame com 1 coluna e 14 linhas:
\describe{
\item{V1}{Expressões}
}
}
}
\source{
Elaborado por Jonathan Ferreira
}
\usage{
dic_4grams
}
\description{
Conjunto de expressões frequentes em decisões judiciais para serem aplicadas sobre tokens para facilitar na análise textual de decisões judiciais.
}
\keyword{datasets}
