% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicionario.R
\name{dicionario.Aplicar}
\alias{dicionario.Aplicar}
\title{Aplica o dicionário de expressões jurídicas}
\usage{
dicionario.Aplicar(token)
}
\arguments{
\item{token}{Objeto do tipo token. Criado com \code{\link[quanteda:tokens]{quanteda::tokens()}}}
}
\value{
Os tokens com as expressões unidas
}
\description{
\code{dicionario.Aplicar()} aplica aos tokens o dicionário de expressões criados para facilitar as análises textuais de decisões judiciais.
O dicionário consta da base de dados do pacote e as atualizações serão informadas no Github
}
