% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicionario.R
\name{dicionario.Criar}
\alias{dicionario.Criar}
\title{Cria um dicionário de expressões a partir de ngrams}
\usage{
dicionario.Criar(token, n = 2, arquivo)
}
\arguments{
\item{token}{Objeto do tipo token. Criado com \code{\link[quanteda:tokens]{quanteda::tokens()}}}

\item{n}{Número de ngrams}

\item{arquivo}{Local para salvar o arquivo. Deve terminar com \code{/}}
}
\value{
Um arquivo csv com as 500 expressões mais frequentes
}
\description{
\code{dicionario.Criar()} cria um dicionário de expressões usando ngrams do pacote \code{quanteda}
Esse dicionário pode ser aplicado depois com a função \code{\link[quanteda:tokens_compound]{quanteda::tokens_compound()}}, permitindo unir palavras que possuem um únicos significado quando em conjunto.
}
