% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicionario.R
\name{dicionario_aplicar}
\alias{dicionario_aplicar}
\title{Aplica dicionários de n-gramas a tokens}
\usage{
dicionario_aplicar(token)
}
\arguments{
\item{token}{Um objeto de tokens, geralmente produzido por funções da biblioteca quanteda.}
}
\value{
Um objeto de tokens modificado após a aplicação dos dicionários de n-gramas.
}
\description{
Esta função aplica uma série de dicionários de n-gramas a um objeto de tokens,
substituindo sequências de tokens que correspondem aos n-gramas nos dicionários
por tokens compostos.
}
\examples{
token <- quanteda::tokens(c("bigram", "analysis", "example"))
token_modificado <- dicionario_aplicar(token)
}
