% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicionario.R
\name{dicionario_criar}
\alias{dicionario_criar}
\title{Cria um dicionário de termos a partir de tokens}
\usage{
dicionario_criar(token, n = 2, arquivo)
}
\arguments{
\item{token}{Um objeto de tokens, normalmente produzido por funções da biblioteca quanteda.}

\item{n}{Um número inteiro especificando o tamanho dos n-gramas. O padrão é 2.}

\item{arquivo}{O caminho do arquivo onde o dicionário será salvo.}
}
\value{
Esta função não retorna um valor diretamente, mas salva um arquivo CSV contendo
o dicionário de termos.
}
\description{
Esta função cria um dicionário de termos a partir de tokens, onde os termos são
gerados a partir de n-gramas, e as frequências dos termos são calculadas a partir
de uma matriz de documentos-frequência.
}
\examples{
token <- quanteda::tokens(c("bigram", "analysis", "example"))
dicionario_criar(token, arquivo = "meu_dicionario.csv")
}
