% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSTF.R
\name{downloadSTF.inicial}
\alias{downloadSTF.inicial}
\title{Baixa a petição inicial do processo de controle concentrado}
\usage{
downloadSTF.inicial(classe, n, UA, arquivo)
}
\arguments{
\item{arquivo}{}
}
\description{
Baixa a petição inicial do processo de controle concentrado
}
