% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSTF.R
\name{downloadSTF.juris}
\alias{downloadSTF.juris}
\title{Baixa as decisões extraídas da página de jurisprudências do STF}
\usage{
downloadSTF.juris(conteudo, arquivo, quantidade)
}
\arguments{
\item{conteudo}{Objeto \code{html} que pode ser obtido com a função \code{\link[=juriSTF.conteudo]{juriSTF.conteudo()}}}

\item{arquivo}{Local onde será salvo as decisões}

\item{quantidade}{Quantidade de decisões que se pretende baixar \(não maior do que as buscadas com a função \code{juriSTF.conteudo}\). Pode ser colocado um valor referente a posição do conteúdo.}
}
\value{
Decisões em \code{.pdf}
}
\description{
\code{downloadSTF.juris} permite baixar as decisões do site de site de \url{jurisprudência.stf.jus.br}.
}
\details{
Igual a função \code{\link[=juriSTF.download]{juriSTF.download()}}
}
