% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrairSTF.R
\name{extrairSTF.relator}
\alias{extrairSTF.relator}
\title{Extrair o relator do processo}
\usage{
extrairSTF.relator(lista, classe, n)
}
\arguments{
\item{lista}{Uma lista}

\item{classe}{Um caracter que indica a classe processual}

\item{n}{Um número que indica o número do processo}
}
\value{
Uma lista com os relatores dos processos requeridos.
}
\description{
\code{extrairSTF.relator()} permite extrair do site do Supremo Tribunal Federal dados de relator dos processos.
}
\examples{
# Extrair os dados da ADPF 800
\dontrun{
extrairSTF.relator(lista, "ADPF", 800, UA)
}

# Extrair os dados da ADI 500 a 600
\dontrun{
extrairSTF.relator(lista, "ADI", 500:600, UA)
}
}
\author{
Essa função teve a colaboração de Gabriel Delias de Sousa Simões
}
