% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisprudencia.R
\name{jurisprudencia_stf}
\alias{jurisprudencia_stf}
\title{Extrair dados do portal de jurisprudência do STF}
\usage{
jurisprudencia_stf(
  busca = " ",
  base = c("acordaos", "decisoes"),
  quantidade = 25
)
}
\arguments{
\item{quantidade}{}
}
\value{
Uma tabela com os dados
}
\description{
A função permite extrair dados de jurisprudência direto do portal do Supremo Tribunal Federal.

Atualmente a busca pode ser feita em duas bases: decisões monocráticas e de acórdãos.

Devido a limitações a quantidade não retorna necessariamente um número exato de decisões. Isto é, se você quer 300 acórdãos e estabelecer que a quantidade é 300, não necessariamente irá retornar 300 acórdãos. Se houver, por exemplo, 800 acórdãos, irá retornar 500 acórdãos, pois a busca sempre retorna a página inteira (com até 250 decisões cada página).

O critério busca pode ser construído utilizando dos operadores \code{AND},  \code{OR}, \code{NO}, \code{""}, \verb{""~}, \code{~}, \code{$},\verb{?} e \verb{()}. Por exemplo, você quer os acórdãos sobre direito ao esquecimento, portanto, deverá pesquisar \code{jurisprudencia_stf('"direito ao esquecimento"', 'acordaos')}. Mas digamos que você queira acórdãos sobre direito à vida e direito à dignidade humana: \code{jurisprudencia_stf('"direito à vida" AND "direito à dignidade humana"', 'acordaos')}
}
\examples{
dados <- jurisprudencia_stf('"direito ao esquecimento"', 'acordaos', 5)
dados

}
