% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisprudencia.R
\name{jurisprudencia_stf_download}
\alias{jurisprudencia_stf_download}
\title{Title}
\usage{
jurisprudencia_stf_download(
  busca = " ",
  base = c("acordaos", "decisoes"),
  quantidade = 25,
  arquivo = "."
)
}
\arguments{
\item{arquivo}{}
}
\description{
Title
}
