% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juriSTF.R
\name{juriSTF.conteudo}
\alias{juriSTF.conteudo}
\title{Pegar o conteúdo da página de jurisprudências do STF}
\usage{
juriSTF.conteudo(busca, quantidade, UA)
}
\arguments{
\item{busca}{Arquivo \code{.json} com os parâmetros de busca}

\item{quantidade}{Número de decisões que devem retornar}

\item{UA}{User-Agent}
}
\value{
Um objeto \code{html} com o conteúdo
}
\description{
\code{juriSTF.conteudo} permite extrair o conteúdo do site de \url{jurisprudência.stf.jus.br}.
}
