% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juriSTF.R
\name{juriSTF.download}
\alias{juriSTF.download}
\title{Baixa as decisões extraídas da página de jurisprudências do STF}
\usage{
juriSTF.download(conteudo, UA, arquivo, quantidade)
}
\arguments{
\item{conteudo}{Objeto \code{html} que pode ser obtido com a função \code{\link[=juriSTF.conteudo]{juriSTF.conteudo()}}}

\item{UA}{User-Agent}

\item{arquivo}{Local onde será salvo as decisões}

\item{quantidade}{Quantidade de decisões que se pretende baixar \(não maior do que as buscadas com a função \code{juriSTF.conteudo}\). Pode ser colocado um valor referente a posição do conteúdo.}
}
\value{
Decisões em \code{.pdf}
}
\description{
\code{juriSTF.download} permite baixar as decisões do site de site de \url{jurisprudência.stf.jus.br}.
}
