% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juriSTF.R
\name{juriSTF.tabela}
\alias{juriSTF.tabela}
\title{Tabelar o conteúdo da página de jurisprudências do STF}
\usage{
juriSTF.tabela(busca, quantidade)
}
\arguments{
\item{busca}{Arquivo \code{.json} com os parâmetros de busca}

\item{quantidade}{Número de decisões que devem retornar}
}
\value{
Uma tabela com o conteúdo
}
\description{
\code{juriSTF.tabela} permite extrair o conteúdo do site de \url{jurisprudência.stf.jus.br} e devolver em uma tabela.
}
