% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lista_partes}
\alias{lista_partes}
\title{Lista para uniformização do nome das partes}
\format{
\subsection{\code{lista_partes}}{

Um character com 842 elementos
}
}
\source{
Elaborado por Jonathan Ferreira
}
\usage{
lista_partes
}
\description{
Objeto do tipo \code{characeter} que contém nome das partes de processos de controle concentrado para serem uniformizadas. Para uniformizar é só utilizar a função da seguinte forma: \verb{[COLUNA/TEXTO QUE SERÁ ALTERADO] |>stringr::str_replace_all(decJ::lista_partes)}
}
\keyword{datasets}
