% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis.R
\name{oasis_to_rayyan}
\alias{oasis_to_rayyan}
\title{Oasis to Rayyan}
\usage{
oasis_to_rayyan(pesquisa)
}
\arguments{
\item{pesquisa}{\code{String} - os descritores da pesquisa. Ver exemplos.}
}
\value{
Um \code{tibble}
}
\description{
Extrai dados do repositório Oasisbr e retorna um objeto compatível com o Rayyan (mas pode ser utilizado para outras coisas também rs)
}
\details{
Por enquanto o único parâmetro que pode ser alteardo é a busca. Estou estudando a alteração dos demais campos de pesquisa.
}
