% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iramuteq.R
\name{plot.iramuteq_stat}
\alias{plot.iramuteq_stat}
\title{Cria um gráfico de estatísticas lexicais produzido no IRaMuTeQ.}
\usage{
\method{plot}{iramuteq_stat}(
  corpus.path,
  corpus.name,
  save.plots = "TRUE",
  plot.path = NULL,
  stat = 1,
  st
)
}
\arguments{
\item{corpus.path}{Local onde está salvo o corpus}

\item{corpus.name}{O nome do corpus}

\item{save.plots}{Lógico.}

\item{plot.path}{Local para salvar o gráfico}

\item{stat}{Padrão é 1. Se você criou mais de uma estatística lexical no IRaMuTeQ indique o número referente. Por exemplo: stat = 3 (se foram criadas três estatísticas e você quer o gráfico da terceira)}

\item{st}{O número de segmentos de textos.}
}
\value{
Um gráfico
}
\description{
Cria um gráfico de estatísticas lexicais produzido no IRaMuTeQ.
}
