% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\name{stf_}
\alias{stf_}
\alias{stf_partes}
\alias{stf_info}
\alias{stf_relator}
\alias{stf_decisoes}
\title{Extrair dados de processos do Supremo Tribunal Federal}
\usage{
stf_()

stf_partes(classe, processo)

stf_info(classe, processo)

stf_relator(classe, processo)

stf_decisoes(classe, processo)
}
\arguments{
\item{classe}{Classe processual. Por exemplo: "ADPF" ou "ADI". Usar abreviatura da classe. Se tiver em dúvida, consultar o site do Supremo.}

\item{processo}{Número dos processos que pretende solicitar. Por exemplo: 500:600 (do 500 ao 600), ou list(500,600,700) que vai buscar o processo 500, 600 e 700}
}
\value{
Data frame
}
\description{
Quatro funções que permitem extrair diferentes dados de processos no Supremo Tribunal Federal.

\code{\link[=stf_partes]{stf_partes()}}: Extrair as partes dos processos

\code{\link[=stf_info]{stf_info()}}: Extrair data de ajuizamento e assuntos do processo

\code{\link[=stf_relator]{stf_relator()}}: Extrair o relator do processo

\code{\link[=stf_decisoes]{stf_decisoes()}}: Extrair as decisões tomadas no processo
}
\examples{
# Retorna 1 processo
tabela_partes <- stf_partes("ADPF", 500)

# Retorna uma sequência de processos
tabela_partes <- stf_partes("ADI", 1500:1505)

# Retorna um conjunto de processos
tabela_partes <- stf_partes("ADO", c(10,15,20))

}
