% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\name{stf_decisoes}
\alias{stf_decisoes}
\title{Extrai informações sobre as decisões de um processo judicial no STF}
\usage{
stf_decisoes(classe, processo)
}
\arguments{
\item{classe}{A classe do processo, em maiúsculas. Por exemplo, "RE", "HC", "ADI", etc.}

\item{processo}{O número do processo ou uma lista de números de processo.}
}
\value{
Um \code{data.frame} contendo informações sobre as decisões do processo, incluindo as seguintes colunas:
\describe{
\item{Classe}{A classe do processo.}
\item{Data}{A data da decisão.}
\item{Nome}{O nome da decisão.}
\item{Julgador}{O julgador responsável pela decisão.}
\item{Decisao}{O conteúdo da decisão.}
}
}
\description{
Esta função extrai informações sobre as decisões de um processo judicial no Supremo Tribunal Federal (STF)
utilizando a API do portal STF. Ela recebe como entrada a classe do processo e o número do processo
e retorna um \code{data.frame} contendo informações sobre as decisões do processo.
}
\details{
Para mais detalhes, acesse:
\code{vignette("familia_stf", package = "decJ")}
}
\examples{
\dontrun{
stf_decisoes("ADPF", "800")
stf_decisoes("ADPF", c("800", "900"))
}
}
\seealso{
Other stf: 
\code{\link{stf_info}()},
\code{\link{stf_inicial}()},
\code{\link{stf_jurisprudencia_download}()},
\code{\link{stf_jurisprudencia}()},
\code{\link{stf_partes}()},
\code{\link{stf_relator}()}
}
\concept{stf}
