% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\name{stf_info}
\alias{stf_info}
\title{Extrai informações sobre um processo judicial no STF}
\usage{
stf_info(classe, processo)
}
\arguments{
\item{classe}{A classe do processo, em maiúsculas. Por exemplo, "RE", "HC", "ADI", etc.}

\item{processo}{O número do processo ou uma lista de números de processo.}
}
\value{
Um \code{data.frame} contendo informações sobre o processo, incluindo as seguintes colunas:
\describe{
\item{Classe}{A classe do processo.}
\item{Ajuizamento}{A data do protocolo (ajuizamento) do processo.}
\item{Assunto}{O assunto do processo.}
}
}
\description{
Esta função extrai informações sobre um processo judicial no Supremo Tribunal Federal (STF)
utilizando a API do portal STF. Ela recebe como entrada a classe do processo e o número do processo
e retorna um \code{data.frame} contendo informações sobre o processo, incluindo a data do protocolo
(ajuizamento) e o assunto do processo.
}
\examples{
stf_info("RE", "12345/2023")
stf_info("ADI", c("54321/2022", "67890/2021"))
}
\seealso{
\code{\link[=stf_partes]{stf_partes()}} \code{\link[=stf_decisoes]{stf_decisoes()}} \code{\link[=stf_relator]{stf_relator()}}
}
