% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\name{stf_inicial}
\alias{stf_inicial}
\title{Baixa a petição inicial de processos do STF}
\usage{
stf_inicial(classe, n, arquivo)
}
\arguments{
\item{classe}{A classe do processo.}

\item{n}{Um vetor contendo os números dos processos.}

\item{arquivo}{O diretório onde os arquivos serão salvos.}
}
\value{
Esta função não retorna nada explicitamente. Ela salva os arquivos PDF da petição inicial dos processos no diretório especificado.
}
\description{
Esta função permite baixar a petição inicial de processos do Supremo Tribunal Federal (STF) com base nos parâmetros fornecidos.
}
\details{
Para mais detalhes, acesse:
\code{vignette("familia_stf", package = "decJ")}
}
\examples{
\dontrun{
# Baixar a petição inicial do processo com número 12345 da classe "RE"
stf_inicial(classe = "ADPF", n = 800, arquivo = ".")
}
}
\seealso{
Other stf: 
\code{\link{stf_decisoes}()},
\code{\link{stf_info}()},
\code{\link{stf_jurisprudencia_download}()},
\code{\link{stf_jurisprudencia}()},
\code{\link{stf_partes}()},
\code{\link{stf_relator}()}
}
\concept{stf}
