% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\name{stf_partes}
\alias{stf_partes}
\title{Extrair partes de um processo no Supremo Tribunal Federal}
\usage{
stf_partes(classe, processo)
}
\arguments{
\item{classe}{Classe processual. Por exemplo: "ADPF" ou "ADI". Usar abreviatura da classe. Se tiver em dúvida, consultar o site do Supremo.}

\item{processo}{Número dos processos que pretende solicitar. Por exemplo: 500:600 (do 500 ao 600), ou list(500,600,700) que vai buscar o processo 500, 600 e 700}
}
\description{
Essa função permite extrair as partes de um processo no Supremo Tribunal Federal, retornando uma tabela com o número do processo, o tipo da parte e o nome da parte.
}
