% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\name{stf_relator}
\alias{stf_relator}
\title{Extrai informações sobre o relator de um processo judicial no STF}
\usage{
stf_relator(classe, processo)
}
\arguments{
\item{classe}{A classe do processo, em maiúsculas. Por exemplo, "RE", "HC", "ADI", etc.}

\item{processo}{O número do processo ou uma lista de números de processo.}
}
\value{
Um \code{data.frame} contendo informações sobre o relator do processo, incluindo as seguintes colunas:
\describe{
\item{Classe}{A classe do processo.}
\item{Relator}{O nome do relator do processo.}
}
}
\description{
Esta função extrai informações sobre o relator de um processo judicial no Supremo Tribunal Federal (STF)
utilizando a API do portal STF. Ela recebe como entrada a classe do processo e o número do processo
e retorna um \code{data.frame} contendo informações sobre o relator do processo.
}
\examples{
stf_relator("RE", "12345/2023")
stf_relator("ADI", c("54321/2022", "67890/2021"))
}
\seealso{
\code{\link[=stf_partes]{stf_partes()}} \code{\link[=stf_info]{stf_info()}} \link{stf_decisoes}
}
