% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitario.R
\name{utilitario_remover_acentos}
\alias{utilitario_remover_acentos}
\title{Remove acentos de um determinado texto}
\usage{
utilitario_remover_acentos(texto)
}
\arguments{
\item{texto}{Um vetor de strings contendo os textos com acentos.}
}
\value{
Um vetor de strings com os acentos removidos.
}
\description{
Esta função remove os acentos de textos.
}
\details{
Esta função substitui os caracteres acentuados por suas
versões não acentuadas. Por exemplo, "Café" é transformado em "Cafe".

Esta função é semelhante à função \code{txt4cs::remove_accent()} do pacote \code{txt4cs},
mas não depende de outros pacotes externos.
}
\examples{
utilitario_remover_acentos(c("Café", "Maçã", "São Paulo"))

}
