% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitario.R
\name{utilitario_remover_cr}
\alias{utilitario_remover_cr}
\title{Remove cabeçalhos e rodapés de texto extraído de documentos PDF.}
\usage{
utilitario_remover_cr(txt_path = ".", rodape = 4, cabecalho = 2)
}
\arguments{
\item{txt_path}{O caminho para a pasta que contém os arquivos PDF. O padrão é o diretório atual.}

\item{rodape}{O número de linhas do rodapé a serem removidas de cada página. O padrão é 4.}

\item{cabecalho}{O número de linhas do cabeçalho a serem removidas de cada página. O padrão é 2.}
}
\value{
Uma tabela com uma coluna de texto e uma coluna de identificação do documento.
}
\description{
Esta função extrai texto de documentos PDF, remove cabeçalhos e rodapés específicos
de cada página e retorna uma tabela com o texto tratado para cada documento.
}
\examples{
utilitario_remover_cr()
utilitario_remover_cr(txt_path = 'caminho/para/pasta', rodape = 3, cabecalho = 1)
}
\seealso{
Other utilitarios: 
\code{\link{utilitario_tscolor}()}
}
\concept{utilitarios}
