% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alepr.R
\name{alepr_projeto}
\alias{alepr_projeto}
\title{Consulta projetos legislativos no site da Assembleia Legislativa do Estado do Paraná}
\usage{
alepr_projeto(
  prop_tipo = "",
  prop_autores = "",
  prop_numero = "",
  prop_ano = "",
  prop_conclusao = "",
  prop_assuntos = "",
  prop_chave = "",
  prop_sumula = "",
  .reportar = TRUE
)
}
\arguments{
\item{prop_tipo}{Tipo da proposição. Pode ser vazio (todos os tipos) ou um número de 1 a 11,
onde:
\itemize{
\item 1: Lei
\item 2: Lei Complementar
\item 3: Resolução
\item 4: Decreto Legislativo
\item 5: Emenda Constitucional
\item 6: Indicação Legislativa
\item 7: Ato
\item 10: Ofício
\item 11: Documento de Designação
\item 12: Portaria
}}

\item{prop_autores}{Autores da proposição.}

\item{prop_numero}{Número da proposição.}

\item{prop_ano}{Ano da proposição.}

\item{prop_conclusao}{Status de conclusão da proposição. Pode ser vazio (todos os tipos) ou um
número de 0 a 5, onde:
\itemize{
\item 0: Sancionada
\item 1: Promulgada
\item 2: Sancionada/Republicada
\item 3: Promulgada/Republicada
\item 4: Revogada
\item 5: Aprovada
}}

\item{prop_assuntos}{Assuntos relacionados à proposição.}

\item{prop_chave}{Chave da proposição.}

\item{prop_sumula}{Sumula da proposição.}

\item{.reportar}{Valor lógico indicando se deve reportar a ação realizada. Padrão é TRUE.}
}
\value{
Retorna uma tabela com os projetos legislativos correspondentes aos parâmetros fornecidos.
Se nenhum projeto for encontrado, retorna NULL.
}
\description{
Esta função realiza uma consulta aos projetos legislativos no site da Assembleia
Legislativa do Estado do Paraná, com base nos parâmetros fornecidos. Os parâmetros
podem incluir o tipo de proposição, autores, número, ano, conclusão, assuntos, chave
e sumula.
}
\examples{
alepr_projeto(prop_tipo = 1, prop_numero = 21890, prop_ano = 2024)

alepr_projeto(prop_tipo = 4, prop_numero = 29, prop_ano = 2020)
}
\seealso{
Other legislativo: 
\code{\link{alers_projeto}()}
}
\concept{legislativo}
