% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrairSTF.R
\name{extrairSTF.info}
\alias{extrairSTF.info}
\title{Extrair data do protocolo e assunto do processo}
\usage{
extrairSTF.info(lista, classe, n, UA)
}
\arguments{
\item{lista}{Uma lista}

\item{classe}{Um caracter que indica a classe processual}

\item{n}{Um número que indica o número do processo}

\item{UA}{User-Agent}
}
\value{
Uma lista com as datas de ajuizamento e assuntos requeridos.
}
\description{
\code{extrairSTF.info()} permite extrair do site do Supremo Tribunal Federal dados de ajuizamento e assunto dos processos.
}
\examples{
# Extrair os dados da ADPF 800
\dontrun{
extrairSTF.info(lista, "ADPF", 800, UA)
}

# Extrair os dados da ADI 500 a 600
\dontrun{
extrairSTF.info(lista, "ADI", 500:600, UA)
}
}
