% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrairSTF.R
\name{extrairSTF.partes}
\alias{extrairSTF.partes}
\title{Extrair partes do processo}
\usage{
extrairSTF.partes(lista, classe, n, UA)
}
\arguments{
\item{lista}{Uma lista}

\item{classe}{Um caracter que indica a classe processual}

\item{n}{Um número que indica o número do processo}

\item{UA}{User-Agent}
}
\value{
Uma lista com os dados das partes requeridos.
}
\description{
\code{extrairSTF.partes()} permite extrair do site do Supremo Tribunal Federal dados das partes dos processos.
}
\examples{
# Extrair os dados da ADPF 800
\dontrun{
extrairSTF.partes(lista, "ADPF", 800, UA)
}

# Extrair os dados da ADI 500 a 600
\dontrun{
extrairSTF.partes(lista, "ADI", 500:600, UA)
}
}
