% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juriSTF.R
\name{juriSTF.tabela}
\alias{juriSTF.tabela}
\title{Tabelar o conteúdo da página de jurisprudências do STF}
\usage{
juriSTF.tabela(busca, quantidade)
}
\arguments{
\item{busca}{Arquivo \code{.json} com os parâmetros de busca}

\item{quantidade}{Número de decisões que devem retornar}
}
\value{
Uma tabela com o conteúdo
}
\description{
\if{html}{\out{
<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-deprecated.svg" alt="Deprecated"/>
}}


Utilize a função \code{\link[=jurisprudencia_stf]{jurisprudencia_stf()}} no lugar.
}
