% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juriSTF.R
\name{juriSTF.tabela250}
\alias{juriSTF.tabela250}
\title{Tabelar o conteúdo grande (+250) da página de jurisprudências do STF}
\usage{
juriSTF.tabela250(busca, quantidade)
}
\arguments{
\item{quantidade}{}
}
\description{
\if{html}{\out{
<img src="https://lifecycle.r-lib.org/articles/figures/lifecycle-deprecated.svg" alt="Deprecated"/>
}}


Utilize a função \code{\link[=jurisprudencia_stf]{jurisprudencia_stf()}} no lugar.
}
