% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis.R
\name{oasis}
\alias{oasis}
\title{Oasis}
\usage{
oasis(pesquisa)
}
\arguments{
\item{pesquisa}{\code{String} - os descritores da pesquisa. Ver exemplos.}
}
\value{
Um \code{tibble}
}
\description{
Extrai dados do repositório Oasisbr e retorna um tibble
}
\details{
Por enquanto o único parâmetro que pode ser alteardo é a busca. Estou estudando a alteração dos demais campos de pesquisa.
}
\examples{
# Pesquisa com apenas uma palavra-chave:
df <- oasis_to_rayyan('judicialização')

# Pesquisa com mais de uma palavra mas que a ordem importa:
df <- oasis_to_rayyan('"judicialização da política"')

# Pesquisa com conjunto(s) de palavras:
df <- oasis_to_rayyan('"judicialização da política" AND "supremo tribunal federal"')
df <- oasis_to_rayyan('("judicialização da política" AND "supremo tribunal federal") AND ("partidos políticos" OR "presidente da república")')
}
