% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis.R
\name{oasis}
\alias{oasis}
\title{Realiza uma pesquisa no OASIS e retorna os resultados}
\usage{
oasis(pesquisa, rayyan = FALSE)
}
\arguments{
\item{pesquisa}{A string contendo o termo de pesquisa a ser usado no OASIS.}

\item{rayyan}{Um valor lógico indicando se os resultados devem ser formatados para uso com a ferramenta Rayyan. O padrão é FALSE.}
}
\value{
Um data frame contendo os metadados dos registros encontrados.
}
\description{
Esta função realiza uma pesquisa no OASIS utilizando a API e retorna os resultados encontrados.
Ela extrai os metadados dos registros encontrados, como título, autores, resumo, palavras-chave, data de publicação, notas e URL.
}
\examples{
# Realizar pesquisa no OASIS e obter os resultados
oasis("direitos humanos")

# Realizar pesquisa no OASIS e formatar os resultados para uso com o Rayyan
oasis("direitos humanos", rayyan = TRUE)
}
