% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\name{stf_jurisprudencia}
\alias{stf_jurisprudencia}
\title{Extrai jurisprudência do Supremo Tribunal Federal (STF)}
\usage{
stf_jurisprudencia(
  busca = NULL,
  classe = NULL,
  base = c("acordaos", "decisoes"),
  quantidade = 25
)
}
\arguments{
\item{busca}{Uma string contendo palavras-chave para a busca de jurisprudência. Se não especificado, a busca será realizada por classe.}

\item{classe}{Uma string contendo a classe de processo para a busca de jurisprudência. Se não especificado, a busca será realizada por palavras-chave.}

\item{base}{Um vetor contendo os tipos de documentos a serem buscados. Pode incluir "acordaos" ou "decisoes".}

\item{quantidade}{O número de resultados desejados. O padrão é 25.}
}
\value{
Um \code{data.frame} contendo as informações da jurisprudência encontrada.
}
\description{
Esta função extrai jurisprudência do Supremo Tribunal Federal (STF) utilizando a API do portal de jurisprudência STF.
Você pode pesquisar por palavras-chave ou por classe de processo.
}
\examples{
\dontrun{
# Buscar jurisprudência por palavras-chave
stf_jurisprudencia(busca = "direitos humanos", base = 'decisoes', quantidade = 10)

# Buscar jurisprudência por classe de processo
stf_jurisprudencia(classe = "ADPF", base = 'acordaos', quantidade = 10)
}
}
\seealso{
Other stf: 
\code{\link{stf_decisoes}()},
\code{\link{stf_info}()},
\code{\link{stf_inicial}()},
\code{\link{stf_jurisprudencia_download}()},
\code{\link{stf_partes}()},
\code{\link{stf_relator}()}
}
\concept{stf}
