% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\name{stf_jurisprudencia_download}
\alias{stf_jurisprudencia_download}
\title{Baixa jurisprudência do Supremo Tribunal Federal (STF)}
\usage{
stf_jurisprudencia_download(
  busca = " ",
  base = c("acordaos", "decisoes"),
  quantidade = 25,
  arquivo = "."
)
}
\arguments{
\item{busca}{Uma string contendo palavras-chave para a busca de jurisprudência. Se não especificado, a busca será realizada por classe.}

\item{base}{Um vetor contendo os tipos de documentos a serem buscados. Pode incluir "acordaos" ou "decisoes".}

\item{quantidade}{O número de documentos a serem baixados. O padrão é 25.}

\item{arquivo}{O diretório onde os arquivos serão salvos. O padrão é o diretório atual.}
}
\description{
Esta função baixa jurisprudência do Supremo Tribunal Federal (STF) utilizando a API do portal de jurisprudência STF.
Você pode pesquisar por palavras-chave ou por classe de processo e especificar a quantidade de documentos a serem baixados.
Os documentos são salvos como arquivos PDF.
}
\examples{
# Baixar jurisprudência por palavras-chave
stf_jurisprudencia_download(busca = "direitos humanos", base = "acordaos", quantidade = 10, arquivo = "~/Downloads/")

# Baixar jurisprudência por classe de processo
stf_jurisprudencia_download(classe = "ADI", quantidade = 10, arquivo = "~/Downloads/")

# Baixar jurisprudência por classe de processo com base específica
stf_jurisprudencia_download(classe = "RE", base = "acordaos", quantidade = 10, arquivo = "~/Downloads/")

}
\seealso{
Other stf: 
\code{\link{stf_decisoes}()},
\code{\link{stf_info}()},
\code{\link{stf_inicial}()},
\code{\link{stf_jurisprudencia}()},
\code{\link{stf_partes}()},
\code{\link{stf_relator}()}
}
\concept{stf}
