% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf.R
\name{stf_partes}
\alias{stf_partes}
\title{Extrai informações sobre as partes de processos judiciais no STF}
\usage{
stf_partes(classe, processo)
}
\arguments{
\item{classe}{A classe do processo, em maiúsculas. Por exemplo, "RE", "HC", "ADI", etc.}

\item{processo}{O número do processo ou uma lista de números de processo.}
}
\value{
Um \code{data.frame} contendo informações sobre as partes envolvidas no processo.
O \code{data.frame} contém as seguintes colunas:
\describe{
\item{Classe}{A classe do processo.}
\item{Tipo}{O tipo da parte (por exemplo, "Autor", "Réu", "Interessado", etc.).}
\item{Parte}{O nome da parte envolvida no processo.}
}
}
\description{
Esta função extrai informações sobre as partes de processos judiciais no Supremo Tribunal Federal (STF)
utilizando a API do portal STF. Ela recebe como entrada a classe do processo e o número do processo
e retorna um \code{data.frame} contendo informações sobre as partes envolvidas no processo.
}
\examples{
stf_partes("RE", "12345/2023")
stf_partes("ADI", c("54321/2022", "67890/2021"))
}
\seealso{
\code{\link[=stf_info]{stf_info()}} \code{\link[=stf_decisoes]{stf_decisoes()}} \code{\link[=stf_relator]{stf_relator()}}
}
