% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stf_partes.R
\name{stf_partes}
\alias{stf_partes}
\title{Extrai informações sobre as partes de processos judiciais no STF}
\usage{
Genérica:
stf_partes(obj, ...)

Classe data.frame
stf_partes(obj, classe, numero)

Classe numeric
stf_partes(obj, classe)
}
\arguments{
\item{obj}{Pode ser um vetor de números, um data frame ou uma lista.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Outros argumentos que devem ser preenchidos conforme a classe do objeto:
\itemize{
\item numeric: você deve informar a classe referente aos números que indicam o processo: Ex: \code{stf_partes(800, "ADPF")} retornará os dados da ADPF 800.
\item data.frame: você deve informar a coluna em que estão os números dos processos e a coluna em que estão as classes dos processos. Ex: \code{stf_partes(tabela, "classes", "numeros")} retornará os dados das ações de numero correspondente aos números da coluna numeros da tabela e a classe da coluna classes da tabela.
\item list: (em desenvolvimento)
}}
}
\value{
Um \code{data.frame} contendo informações sobre as partes envolvidas no processo.
O \code{data.frame} contém as seguintes colunas:
\describe{
\item{Classe}{A classe do processo.}
\item{Tipo}{O tipo da parte (por exemplo, "Autor", "Réu", "Interessado", etc.).}
\item{Parte}{O nome da parte envolvida no processo.}
}
}
\description{
Esta função extrai informações sobre as partes de processos judiciais no Supremo Tribunal Federal (STF)
utilizando a API do portal STF. Ela recebe como entrada a classe do processo e o número do processo
e retorna um \code{data.frame} contendo informações sobre as partes envolvidas no processo.
}
\details{
Para mais detalhes, acesse:
\code{vignette("familia_stf", package = "decJ")}
}
\examples{
\dontrun{
stf_partes("ADPF", "800")
stf_partes("ADPF", c("800", "900"))
}
}
\seealso{
Other stf: 
\code{\link{stf_decisoes}()},
\code{\link{stf_info}()},
\code{\link{stf_inicial}()},
\code{\link{stf_jurisprudencia_download}()},
\code{\link{stf_jurisprudencia}()},
\code{\link{stf_relator}()}
}
\concept{stf}
